celery_worker/requirements.txt
celery==5.4.0
redis==5.0.7
clickhouse-connect==0.7.16
prometheus-client==0.20.0
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-instrumentation-celery==0.47b0
